+++ src/osgPlugins/txp/trpage_writebuf.cpp
@@ -332,14 +332,14 @@ bool trpgwAppFile::Append(const trpgMemWriteBuffer *buf1,const trpgMemWriteBuffe
     // Write the data out
     const char *data = buf1->getData();
     int len = buf1->length();
-    if (fwrite(data,sizeof(char),len,fp) != len) {
+    if (fwrite(data,sizeof(char),len,fp) != static_cast<unsigned int>(len)) {
         valid = false;
         return false;
     }
     if (buf2) {
         data = buf2->getData();
         len = buf2->length();
-        if (fwrite(data,sizeof(char),len,fp) != len) {
+        if (fwrite(data,sizeof(char),len,fp) != static_cast<unsigned int>(len)) {
             valid = false;
             return false;
         }
@@ -362,7 +362,7 @@ bool trpgwAppFile::Append(const char *data,int size)
     }
 
     // Write the data out
-    if (fwrite(data,sizeof(char),size,fp) != size) {
+    if (fwrite(data,sizeof(char),size,fp) != static_cast<unsigned int>(size)) {
         valid = false;
         return false;
     }
